+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/metamodel/AbstractMethodPluralAttribute.java
@@ -52,12 +52,14 @@ public abstract class AbstractMethodPluralAttribute<X, C, Y> extends AbstractMet
     private final boolean updateCascaded;
     private final boolean deleteCascaded;
     private final boolean orphanRemoval;
+    private final Set<Type<?>> readOnlySubtypes;
     private final Set<Type<?>> persistSubtypes;
     private final Set<Type<?>> updateSubtypes;
     private final Set<Class<?>> allowedSubtypes;
     private final Map<ManagedViewType<? extends Y>, String> elementInheritanceSubtypes;
     private final boolean sorted;
     private final boolean ordered;
+    private final boolean forcedUnique;
     private final Class<Comparator<Object>> comparatorClass;
     private final Comparator<Object> comparator;
 
@@ -89,6 +91,8 @@ public abstract class AbstractMethodPluralAttribute<X, C, Y> extends AbstractMet
         }
         boolean definesDeleteCascading = mapping.getCascadeTypes().contains(CascadeType.DELETE);
         boolean allowsDeleteCascading = updatable || mapping.getCascadeTypes().contains(CascadeType.AUTO);
+
+        this.readOnlySubtypes = (Set<Type<?>>) (Set) mapping.getReadOnlySubtypes(context);
 
         if (updatable) {
             this.persistSubtypes = determinePersistSubtypeSet(elementType, mapping.getCascadeSubtypes(context), mapping.getCascadePersistSubtypes(context), context);
@@ -177,6 +181,7 @@ public abstract class AbstractMethodPluralAttribute<X, C, Y> extends AbstractMet
         this.sorted = mapping.isSorted();
         
         this.ordered = mapping.getContainerBehavior() == AttributeMapping.ContainerBehavior.ORDERED;
+        this.forcedUnique = mapping.isForceUniqueness() || determineForcedUnique(context);
         this.comparatorClass = (Class<Comparator<Object>>) mapping.getComparatorClass();
         this.comparator = MetamodelUtils.getComparator(comparatorClass);
     }
@@ -256,6 +261,11 @@ public abstract class AbstractMethodPluralAttribute<X, C, Y> extends AbstractMet
     }
 
     @Override
+    public Set<Type<?>> getReadOnlyAllowedSubtypes() {
+        return readOnlySubtypes;
+    }
+
+    @Override
     public Set<Type<?>> getPersistCascadeAllowedSubtypes() {
         return persistSubtypes;
     }
@@ -308,6 +318,11 @@ public abstract class AbstractMethodPluralAttribute<X, C, Y> extends AbstractMet
     @Override
     public boolean isOrdered() {
         return ordered;
+    }
+
+    @Override
+    public boolean isForcedUnique() {
+        return forcedUnique;
     }
 
     @Override