+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/metamodel/AbstractParameterPluralAttribute.java
@@ -35,6 +35,7 @@ public abstract class AbstractParameterPluralAttribute<X, C, Y> extends Abstract
     private final Map<ManagedViewType<? extends Y>, String> elementInheritanceSubtypes;
     private final boolean sorted;
     private final boolean ordered;
+    private final boolean forcedUnique;
     private final Class<Comparator<Object>> comparatorClass;
     private final Comparator<Object> comparator;
 
@@ -46,6 +47,7 @@ public abstract class AbstractParameterPluralAttribute<X, C, Y> extends Abstract
         this.sorted = mapping.isSorted();
 
         this.ordered = mapping.getContainerBehavior() == AttributeMapping.ContainerBehavior.ORDERED;
+        this.forcedUnique = mapping.isForceUniqueness() || determineForcedUnique(context);
         this.comparatorClass = (Class<Comparator<Object>>) mapping.getComparatorClass();
         this.comparator = MetamodelUtils.getComparator(comparatorClass);
     }
@@ -88,6 +90,11 @@ public abstract class AbstractParameterPluralAttribute<X, C, Y> extends Abstract
     @Override
     public boolean isOrdered() {
         return ordered;
+    }
+
+    @Override
+    public boolean isForcedUnique() {
+        return forcedUnique;
     }
     
     @Override